Move OpRequest, ReqOpType, and RequestKey
