+++ automerge-protocol/src/serde_impls/mod.rs
@@ -4,3 +4,8 @@ mod object_id;
 mod element_id;
 mod key;
 mod value;
+mod request_key;
+
+pub(crate) fn make_false() -> bool {
+    return false
+}
