+++ automerge-protocol/src/serde_impls/request_key.rs
@@ -1,0 +1,34 @@
+use crate::RequestKey;
+use serde::{de, Deserialize, Deserializer, de::Visitor};
+
+impl<'de> Deserialize<'de> for RequestKey {
+    fn deserialize<D>(deserializer: D) -> Result<Self, D::Error>
+    where
+        D: Deserializer<'de>,
+    {
+        struct RequestKeyVisitor;
+        impl<'de> Visitor<'de> for RequestKeyVisitor {
+            type Value = RequestKey;
+
+            fn expecting(&self, formatter: &mut std::fmt::Formatter) -> std::fmt::Result {
+                formatter.write_str("a number or string")
+            }
+
+            fn visit_u64<E>(self, value: u64) -> Result<RequestKey, E>
+            where
+                E: de::Error,
+            {
+                Ok(RequestKey::Num(value))
+            }
+
+            fn visit_str<E>(self, value: &str) -> Result<RequestKey, E>
+            where
+                E: de::Error,
+            {
+                Ok(RequestKey::Str(value.to_string()))
+            }
+        }
+        deserializer.deserialize_any(RequestKeyVisitor)
+    }
+}
+
